# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NoteFlow - multi-user notes service
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  userId: string;
  /** @nullable */
  userEmail: string | null;
  /** @nullable */
  userFirstName?: string | null;
  isPinned: boolean;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  content: string;
  tags?: string[];
  isPinned?: boolean;
}

export interface NoteUpdate {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  content?: string;
  tags?: string[];
  isPinned?: boolean;
}

export interface NoteListResponse {
  notes: Note[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface NoteStats {
  total: number;
  pinned: number;
  totalTags: number;
  recentCount: number;
  topTags?: TagCount[];
}

export type ListNotesParams = {
  /**
   * Search query for title or content
   */
  search?: string;
  /**
   * Page number (1-based)
   */
  page?: number;
  /**
   * Notes per page
   */
  limit?: number;
  /**
   * Filter by tag
   */
  tag?: string;
  /**
   * Filter by pinned status
   */
  pinned?: boolean;
};

export type AdminListNotesParams = {
  search?: string;
  page?: number;
  limit?: number;
  /**
   * Filter by user ID
   */
  userId?: string;
};