# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for NoteFlow - multi-user notes service
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: notes
    description: Note operations
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes (with search and pagination)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search query for title or content
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (1-based)
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
          description: Notes per page
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag
        - name: pinned
          in: query
          schema:
            type: boolean
          description: Filter by pinned status
      responses:
        "200":
          description: Paginated list of notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListResponse"
        "401":
          description: Unauthorized
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
        "401":
          description: Unauthorized

  /notes/stats:
    get:
      operationId: getNoteStats
      tags: [notes]
      summary: Get note statistics for the current user
      responses:
        "200":
          description: Note statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteStats"
        "401":
          description: Unauthorized

  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not owner)
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /notes/{id}/pin:
    patch:
      operationId: togglePin
      tags: [notes]
      summary: Toggle pinned status of a note (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /admin/notes:
    get:
      operationId: adminListNotes
      tags: [admin]
      summary: Admin - list all notes across all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID
      responses:
        "200":
          description: All notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)

  /admin/notes/{id}:
    delete:
      operationId: adminDeleteNote
      tags: [admin]
      summary: Admin - delete any note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Note:
      type: object
      required: [id, title, content, userId, userEmail, isPinned, tags, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        userId:
          type: string
        userEmail:
          type: ["string", "null"]
        userFirstName:
          type: ["string", "null"]
        isPinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    NoteInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean

    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean

    NoteListResponse:
      type: object
      required: [notes, total, page, limit, totalPages]
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    NoteStats:
      type: object
      required: [total, pinned, totalTags, recentCount]
      properties:
        total:
          type: integer
        pinned:
          type: integer
        totalTags:
          type: integer
        recentCount:
          type: integer
        topTags:
          type: array
          items:
            $ref: "#/components/schemas/TagCount"

    TagCount:
      type: object
      required: [tag, count]
      properties:
        tag:
          type: string
        count:
          type: integer