# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NoteFlow - multi-user notes service
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  content: string;
  userId: string;
  /** @nullable */
  userEmail: string | null;
  /** @nullable */
  userFirstName?: string | null;
  isPinned: boolean;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}