# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NoteFlow - multi-user notes service
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  content: string;
  tags?: string[];
  isPinned?: boolean;
}