# lib/api-zod/src/generated/types/noteListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NoteFlow - multi-user notes service
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from "./note";

export interface NoteListResponse {
  notes: Note[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}