# lib/api-zod/src/generated/types/noteStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NoteFlow - multi-user notes service
 * OpenAPI spec version: 0.1.0
 */
import type { TagCount } from "./tagCount";

export interface NoteStats {
  total: number;
  pinned: number;
  totalTags: number;
  recentCount: number;
  topTags?: TagCount[];
}